refactor xtrellis
